+++ packages/FileEditors/src/package-api.ts
@@ -32,6 +32,10 @@ export namespace funcs {
     return await grok.functions.call('FileEditors:PreviewRtf', { file });
   }
 
+  export async function renderMath(latex: string , displayMode: boolean ): Promise<string> {
+    return await grok.functions.call('FileEditors:RenderMath', { latex, displayMode });
+  }
+
   export async function previewTex(file: DG.FileInfo ): Promise<DG.View> {
     return await grok.functions.call('FileEditors:PreviewTex', { file });
   }